chore: moved service structures to separate file
